Project/Socialmediafeeder/requirements.txt
aniso8601==9.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==7.1.2
configure==0.5
cryptography==3.4.7
facebook-sdk==3.1.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
flask-swagger-ui==3.36.0
idna==3.3
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
oauthlib==3.1.1
pycparser==2.20
PyJWT==2.1.0
pymongo==3.11.3
python-dotenv==0.19.0
pytz==2021.1
PyYAML==6.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.15.0
tweepy==4.4.0
twilio==7.3.2
urllib3==1.26.7
Werkzeug==1.0.1
